test: adapt NativeConnection lifecycle tests for Cloud
